+++ p2p/dht/dht.go
@@ -2,30 +2,53 @@ package dht
 
 import (
 	"context"
+	"crypto/sha256"
+	"encoding/hex"
+	"fmt"
+	"math/rand"
+	"time"
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/glog"
 	"github.com/invin/kkchain/p2p"
-	"github.com/invin/kkchain/p2p/dht/pb"
-	"github.com/libp2p/go-libp2p-crypto"
 )
 
 const (
 	protocolDHT = "/kkchain/p2p/dht/1.0.0"
 )
 
+const (
+	DefaultSyncTableInterval   = 10 * time.Second
+	DefaultSaveTableInterval   = 2 * time.Minute
+	DefaultSeedMinTableTime    = 6 * time.Minute
+	DefaultMaxPeersCountToSync = 6
+)
+
+type DHTConfig struct {
+	BucketSize      int
+	RoutingTableDir string
+	BootstrapNodes  []string
+}
+
 // DHT implements a Distributed Hash Table for p2p
 type DHT struct {
 	// self
 	host p2p.Host
 
-	quitCh         chan bool
-	table          *RoutingTable
-	store          *PeerStore
-	config         *p2p.Config
-	self           PeerID
-	selfPrivateKey crypto.PrivKey
-	recvCh         chan interface{}
+	quitCh chan bool
+	table  *RoutingTable
+	store  *PeerStore
+	config *DHTConfig
+	self   PeerID
+	recvCh chan interface{}
+}
+
+func DefaultConfig() *DHTConfig {
+	return &DHTConfig{
+		BucketSize:      BucketSize,
+		RoutingTableDir: "",
+		BootstrapNodes:  []string{},
+	}
 }
 
 func (dht *DHT) GetRecvchan() chan interface{} {
@@ -33,7 +56,7 @@ func (dht *DHT) GetRecvchan() chan interface{} {
 }
 
 // NewDHT creates a new DHT object with the given peer as as the 'local' host
-func NewDHT(config *Config, host p2p.Host) *DHT {
+func NewDHT(config *DHTConfig, host p2p.Host) *DHT {
 
 	// If no node database was given, use an in-memory one
 	db, err := newPeerStore(config.RoutingTableDir)
@@ -41,19 +64,16 @@ func NewDHT(config *Config, host p2p.Host) *DHT {
 		return nil
 	}
 
-	self, privateKey, err := selfPeerID(config)
-	if err != nil {
-		return nil
-	}
+	self := CreateID(host.ID().Address, host.ID().PublicKey)
 
 	dht := &DHT{
-		quitCh:         make(chan bool),
-		config:         config,
-		selfPrivateKey: privateKey,
-		self:           *self,
-		table:          CreateRoutingTable(*self),
-		store:          db,
-		recvCh:         make(chan interface{}),
+		quitCh: make(chan bool),
+		config: config,
+		//selfPrivateKey: privateKey,
+		self:   self,
+		table:  CreateRoutingTable(self),
+		store:  db,
+		recvCh: make(chan interface{}),
 	}
 
 	dht.host = host
@@ -65,44 +85,16 @@ func NewDHT(config *Config, host p2p.Host) *DHT {
 	return dht
 }
 
-//initNetwork init network and related config before startup
-//func initNetwork(config *p2p.Config, dht *DHT) {
-//	//TODO: new server & and set to dht
-//
-//	//dht.host = impl.NewHost(dht.self.ID)
-//}
-
-// selfPeerID get local peer info
-func selfPeerID(config *p2p.Config) (*PeerID, crypto.PrivKey, error) {
-
-	peerKey, err := LoadNetworkKeyFromFileOrCreateNew(config.PrivateKeyPath)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	pubk, err := peerKey.GetPublic().Bytes()
-	if err != nil {
-		return nil, nil, err
-	}
-	id := CreateID(config.Listen, pubk)
-
-	return &id, peerKey, nil
-}
-
 func (dht *DHT) Self() PeerID {
 	return dht.self
-}
-
-func (dht *DHT) SetHost(host p2p.Host) {
-	dht.host = host
 }
 
 // handleNewStream handles messages within the stream
 func (dht *DHT) handleNewStream(s p2p.Stream, msg proto.Message) {
 	// check message type
 	switch message := msg.(type) {
-	case *pb.Message:
-		dht.handleMessage(s, message)
+	case *Message:
+		go dht.handleMessage(s, message)
 	default:
 		s.Reset()
 		glog.Errorf("unexpected message: %v", msg)
@@ -110,7 +102,7 @@ func (dht *DHT) handleNewStream(s p2p.Stream, msg proto.Message) {
 }
 
 // handleMessage handles messsage
-func (dht *DHT) handleMessage(s p2p.Stream, msg *pb.Message) {
+func (dht *DHT) handleMessage(s p2p.Stream, msg *Message) {
 	// get handler
 	handler := dht.handlerForMsgType(msg.GetType())
 	if handler == nil {
@@ -137,5 +129,145 @@ func (dht *DHT) handleMessage(s p2p.Stream, msg *pb.Message) {
 		s.Reset()
 		glog.Errorf("send response error: %s", err)
 		return
+	}
+}
+
+func (dht *DHT) Start() {
+	fmt.Println("start sync loop.....")
+	go dht.syncLoop()
+	go dht.waitReceive()
+}
+
+func (dht *DHT) Stop() {
+	fmt.Println("stopping sync loop.....")
+	dht.quitCh <- true
+}
+
+func (dht *DHT) syncLoop() {
+
+	//TODO: config timer
+	syncLoopTicker := time.NewTicker(DefaultSyncTableInterval)
+	defer syncLoopTicker.Stop()
+
+	saveTableToStore := time.NewTicker(DefaultSaveTableInterval)
+	defer saveTableToStore.Stop()
+
+	for {
+		select {
+		case <-dht.quitCh:
+			fmt.Println("stopped sync loop")
+			return
+		case <-syncLoopTicker.C:
+			dht.SyncRouteTable()
+		case <-saveTableToStore.C:
+			dht.saveTableToStore()
+		}
+	}
+}
+
+func (dht *DHT) waitReceive() {
+	for {
+		select {
+		case msg := <-dht.recvCh:
+			fmt.Println(msg)
+		}
+	}
+}
+
+func (dht *DHT) AddPeer(peer PeerID) {
+
+	//dht.store.Update(&peer)
+	peer.addTime = time.Now()
+	dht.table.Update(peer)
+
+}
+
+func (dht *DHT) RemovePeer(peer PeerID) {
+
+	dht.store.Delete(&peer)
+	dht.table.RemovePeer(peer)
+
+}
+
+//FindTargetNeighbours searches target's neighbours from given PeerID
+func (dht *DHT) FindTargetNeighbours(target []byte, peer PeerID) {
+
+	if peer.Equals(dht.self) {
+		return
+	}
+
+	conn, err := dht.host.GetConnection(peer.ID)
+	if err != nil {
+		return
+	}
+
+	//TODO: dial remote peer???
+	if conn == nil {
+		err = dht.host.Connect(peer.ID.Address)
+		if err != nil {
+			return
+		}
+	}
+
+	//send find neighbours request to peer
+	pmes := NewMessage(Message_FIND_NODE, hex.EncodeToString(target))
+	dht.host.SendMsg(conn, protocolDHT, pmes)
+}
+
+// RandomTargetID generate random peer id for query target
+func RandomTargetID() []byte {
+	id := make([]byte, 32)
+	rand.Read(id)
+	return sha256.New().Sum(id)
+}
+
+// SyncRouteTable sync route table.
+func (dht *DHT) SyncRouteTable() {
+	fmt.Println("timer trigger")
+	fmt.Printf("table size: %d\n", len(dht.table.GetPeers()))
+
+	target := RandomTargetID()
+	syncedPeers := make(map[string]bool)
+
+	// sync with seed nodes.
+	for _, addr := range dht.config.BootstrapNodes {
+		pid, err := ParsePeerAddr(addr)
+		if err != nil {
+			continue
+		}
+
+		dht.FindTargetNeighbours(target, *pid)
+		syncedPeers[hex.EncodeToString(pid.Hash)] = true
+	}
+
+	// random peer selection.
+	peers := dht.table.GetPeers()
+	peersCount := len(peers)
+	if peersCount <= 1 {
+		return
+	}
+
+	peersCountToSync := DefaultMaxPeersCountToSync
+	if peersCount < peersCountToSync {
+		peersCountToSync = peersCount
+	}
+
+	for i := 0; i < peersCountToSync; i++ {
+		pid := peers[i]
+		if syncedPeers[hex.EncodeToString(pid.Hash)] == false {
+			dht.FindTargetNeighbours(target, pid)
+			syncedPeers[hex.EncodeToString(pid.Hash)] = true
+		}
+	}
+}
+
+// saveTableToStore save peer to db
+func (dht *DHT) saveTableToStore() {
+	peers := dht.table.GetPeers()
+	now := time.Now()
+	for _, v := range peers {
+		if now.Sub(v.addTime) > DefaultSeedMinTableTime {
+			dht.store.Update(&v)
+		}
 	}
 }
