+++ p2p/dht/message_test.go
@@ -1,0 +1,17 @@
+package dht
+
+import (
+	"testing"
+)
+
+func TestNewMessage(t *testing.T) {
+	t.Parallel()
+}
+
+func TestPeerIDToPBPeer(t *testing.T) {
+	t.Parallel()
+}
+
+func TestPBPeerToPeerID(t *testing.T) {
+	t.Parallel()
+}
